Fixed 500 error on delete an empty file.

See merge request assisken/zhipa!69

+++ main/models.py
@@ -45,8 +45,9 @@ class File(models.Model):
 
     def delete(self, using=None, keep_parents=False):
         super().delete()
-        os.remove(self.file.path)
-        os.removedirs(os.path.dirname(self.file.path))
+        if self.file:
+            os.remove(self.file.path)
+            os.removedirs(os.path.dirname(self.file.path))
 
 
 class News(models.Model):
